producer-api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
kafka-python==2.0.2
redis==5.0.1
